fix: move hasYarn to its own module
