+++ src/download-package.js
@@ -26,6 +26,9 @@ async function downloadPackage(name, url, range) {
 
   let newTmpDir = await createTmpDir();
   let { stdout } = await spawn('npm', ['install', url], { cwd: newTmpDir });
+  // if (!name) {
+  //   name = stdout.match(/^\+ (.*)@\d\.\d\.\d.*$/m)[1];
+  // }
   let arg = stdout.match(/^\+ (.*)$/m)[1];
   let parsed = npa(arg);
   name = parsed.name;
